Move pointer_category into hpx::traits
