+++ hpx/traits.hpp
@@ -229,6 +229,16 @@ namespace hpx { namespace traits
 
     template <typename T, typename Enable = void>
     struct is_executor_parameters;
+
+    ///////////////////////////////////////////////////////////////////////////
+    struct general_pointer_tag {};
+
+#if defined(HPX_HAVE_CXX11_IS_TRIVIALLY_COPYABLE)
+    struct trivially_copyable_pointer_tag : general_pointer_tag {};
+#endif
+
+    template <typename Source, typename Dest>
+    general_pointer_tag get_pointer_category(Source const&, Dest const&);
 }}
 
 #endif
